Allow skipping fields that may cause conflicts (#25)
The network segmentation ID as well as the port mac addresses may be in-use on the destination cloud.

As such, we'll skip those by default. This can be configured through:

* `preserve_network_segmentation_id`
* `preserve_port_mac_address`

While at it, we're rearranging the config option definitions and fixing a FIP migration handler issue, which is now failing if the FIP is not attached.